+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ### master
+
+### v0.3.0, 2014-08-29
 - bugfix: when top is running the pane `$PWD` can't be saved. This was causing
   issues during the restore and is now fixed.
 - restoring sessions multiple times messes up the whole environment - new panes